Move resolvers to own folder
