+++ packages/hyperdrive-appconfig/src/appconfig/AppConfig.ts
@@ -3,7 +3,7 @@ import { ChainConfig, ChainId } from "src/chains/chains";
 import { HyperdriveConfig } from "src/hyperdrives/HyperdriveConfig";
 import { AnyRewardKey } from "src/hyperdrives/rewards";
 import { protocols } from "src/protocols";
-import { RewardResolverKey } from "src/rewards/rewards";
+import { RewardResolverKey } from "src/rewards/resolvers";
 import { TokenConfig } from "src/tokens/types";
 import { yieldSources } from "src/yieldSources/yieldSources";
 import { ZapConfig } from "src/zaps/ZapsConfig";
