+++ packages/hyperdrive-appconfig/src/rewards/resolvers/etherfi.ts
